Working first implementation of file hashing.

+++ src/main/java/de/weltraumschaf/maconha/repo/BaseRepo.java
@@ -9,6 +9,6 @@ import org.springframework.data.repository.NoRepositoryBean;
  * @param <E> type of entity
  */
 @NoRepositoryBean
-interface BaseRepo<E> extends JpaRepository<E, Integer> {
+interface BaseRepo<E> extends JpaRepository<E, Long> {
 
 }
